+++ datasets/retina_mnist.py
@@ -13,7 +13,7 @@ class RetinaMNISTDataset(Dataset):
     train_dataset: RetinaMNIST
     test_dataset: RetinaMNIST
 
-    hold_out_size: int
+    hold_out_size: float
     X_train: ndarray
     y_train: ndarray
     X_hold_out: ndarray
@@ -21,7 +21,7 @@ class RetinaMNISTDataset(Dataset):
     X_test: ndarray
     y_test: ndarray
 
-    def __init__(self, hold_out_size: int, root_path: str):
+    def __init__(self, hold_out_size: float, root_path: str):
         self.hold_out_size = hold_out_size
         self.train_dataset = RetinaMNIST(
             root=root_path,
